time: correct rounding in update_wall_time()
The update to xtime.tv_nsec in update_wall_time() rounds up to the nearest
nanosecond.  However, this code incorrectly rounds even when there is no
remainder resulting in a very slight dilation in wall-time.  This occurs
specifically in the case where timekeeper is ns accurate, e.g.
timekeeper.shift == 0

Signed-off-by: Paul Turner <pjt@google.com>